requirements.txt
Flask>=2.3.0
openai>=0.28.0
requests>=2.31.0
Pillow>=10.0.0
reportlab>=4.0.0
azure-cognitiveservices-speech>=1.32.0
python-dotenv>=1.0.0
google-genai>=1.29.0